pyproject.toml
[tool.poetry]
name = "autotagger"
version = "1.0"
description = "Danbooru autotagger"
authors = ["evazion <noizave@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
fastbook = "^0.0.26"
Flask = "^2.1.2"
python-dotenv = "^0.20.0"
click = "^8.1.3"
ipywidgets = "^7.7.0"
timm = "^0.5.4"
scipy = "^1.8.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
